Update to Timing & Secret (#4)
Opportunity can be created by a new change in the landscape (market, industry, culture, tech) or by looking at things that have not changed in a while. Things that have not changed highlight an area that is either archaic (and needs fundamental change) or a great source of consistent need.